Improve devops and add documentation
Since many people are starting to use this library, it's a good idea to
improve it's structure. I've gone ahead and initialized the project with
NPM. This lets us create scripts for minifying and generating the
documentation, for example.

I've also written basic instructions in the CONTRIBUTING.md file, but
other than that, you can enable GitHub Pages to host the documentation
from the docs folder.